+++ samplesheets/rules.py
@@ -23,30 +23,29 @@ from projectroles import rules as pr_rules  # To access common predicates
 rules.add_perm(
     'samplesheets.view_sheet',
     rules.is_superuser | pr_rules.is_project_owner |
-    pr_rules.is_project_delegate | pr_rules.is_project_staff |
-    pr_rules.is_project_contributor | pr_rules.is_project_guest)
+    pr_rules.is_project_delegate | pr_rules.is_project_contributor |
+    pr_rules.is_project_guest)
 
 # Allow creating, importing or modifying the project's JSON sample sheet
 rules.add_perm(
     'samplesheets.edit_sheet',
     rules.is_superuser | pr_rules.is_project_owner |
-    pr_rules.is_project_delegate | pr_rules.is_project_staff)
+    pr_rules.is_project_delegate)
 
 # Allow exporting a JSON sample sheet from project
 rules.add_perm(
     'samplesheets.export_sheet',
     rules.is_superuser | pr_rules.is_project_owner |
-    pr_rules.is_project_delegate | pr_rules.is_project_staff |
-    pr_rules.is_project_contributor)
+    pr_rules.is_project_delegate | pr_rules.is_project_contributor)
 
 # Allow creating directory structure in iRODS
 rules.add_perm(
     'samplesheets.create_dirs',
     rules.is_superuser | pr_rules.is_project_owner |
-    pr_rules.is_project_delegate | pr_rules.is_project_staff)
+    pr_rules.is_project_delegate)
 
 # Allow deleting the project sample sheet
 rules.add_perm(
     'samplesheets.delete_sheet',
     rules.is_superuser | pr_rules.is_project_owner |
-    pr_rules.is_project_delegate | pr_rules.is_project_staff)
+    pr_rules.is_project_delegate)