+++ timeline/rules.py
@@ -25,8 +25,8 @@ from projectroles import rules as pr_rules  # To access common predicates
 rules.add_perm(
     'timeline.view_timeline',
     rules.is_superuser | pr_rules.is_project_owner |
-    pr_rules.is_project_delegate | pr_rules.is_project_staff |
-    pr_rules.is_project_contributor | pr_rules.is_project_guest)
+    pr_rules.is_project_delegate | pr_rules.is_project_contributor |
+    pr_rules.is_project_guest)
 
 # Allow viewing classified event
 rules.add_perm(