+++ timeline/tests/test_ui.py
@@ -18,7 +18,6 @@ PROJECT_ROLE_OWNER = OMICS_CONSTANTS['PROJECT_ROLE_OWNER']
 PROJECT_ROLE_DELEGATE = OMICS_CONSTANTS['PROJECT_ROLE_DELEGATE']
 PROJECT_ROLE_CONTRIBUTOR = OMICS_CONSTANTS['PROJECT_ROLE_CONTRIBUTOR']
 PROJECT_ROLE_GUEST = OMICS_CONSTANTS['PROJECT_ROLE_GUEST']
-PROJECT_ROLE_STAFF = OMICS_CONSTANTS['PROJECT_ROLE_STAFF']
 PROJECT_TYPE_CATEGORY = OMICS_CONSTANTS['PROJECT_TYPE_CATEGORY']
 PROJECT_TYPE_PROJECT = OMICS_CONSTANTS['PROJECT_TYPE_PROJECT']
 
@@ -58,7 +57,6 @@ class TestListView(
             (self.superuser, 2),
             (self.as_owner.user, 2),
             (self.as_delegate.user, 2),
-            (self.as_staff.user, 1),
             (self.as_contributor.user, 1),
             (self.as_guest.user, 1)]
 
@@ -85,7 +83,6 @@ class TestListView(
             (self.superuser, 2),
             (self.as_owner.user, 2),
             (self.as_delegate.user, 2),
-            (self.as_staff.user, 1),
             (self.as_contributor.user, 1),
             (self.as_guest.user, 1)]
 
@@ -103,7 +100,6 @@ class TestListView(
             (self.superuser, 1),
             (self.as_owner.user, 1),
             (self.as_delegate.user, 1),
-            (self.as_staff.user, 1),
             (self.as_contributor.user, 1),
             (self.as_guest.user, 1)]